+++ AJData/AJDataTests/Repositories/Forecast/ForecastRepositoryTests.swift
@@ -19,7 +19,7 @@ import Combine
 
 final class ForecastRepositoryTests: XCTestCase {
 
-    private let timeout:Double = 5
+    private let timeout:Float = 5
     private var repository: ForecastRepositoryType!
     private var cancellables: Set<AnyCancellable>!
 
@@ -57,8 +57,8 @@ final class ForecastRepositoryTests: XCTestCase {
 
         //arrange
         let expectation = self.expectation(description: #function)
-        let latitude: Double = 40.64
-        let longitude: Double = -8.64
+        let latitude: Float = 40.64
+        let longitude: Float = -8.64
         let cityExpected = "Aveiro"
         let cityIdExpected = 2742611
